pyproject.toml
[project]
name = "aigenius-ticketing"
version = "1.0.0"
description = "AI-powered Customer Support Ticketing System - SLA Monitoring, Intelligent Classification, and RAG-based Response Generation"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "AIGenius Team"}
]
dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    # Database
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    # Background Tasks
    "apscheduler>=3.10.4",
    # LLM & RAG - Z.AI SDK (GLM 4.7)
    "zai-sdk>=0.1.0",
    # Vector Store - Milvus
    "pymilvus>=2.3.0",
    "milvus>=2.3.0",
    # HTTP Client
    "httpx>=0.26.0",
    # Configuration & Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "watchdog>=4.0.0",
    # Metrics
    "prometheus-client>=0.19.0",
    # Logging
    "python-json-logger>=2.0.0",
    "sniffio>=1.3.1",
    "greenlet>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "httpx>=0.26.0",  # For TestClient
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "httpx>=0.26.0",
]

[tool.uv.sources]
# Pin specific versions if needed for stability
# zai-sdk = { git = "https://github.com/zhipuai/zai-sdk-python.git" }

# ========== Tool Configurations ==========

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line length handled by formatter

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]
